% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{execute_rename_plan}
\alias{execute_rename_plan}
\title{Execute a renaming plan for camera trap images}
\usage{
execute_rename_plan(plan_tbl, overwrite = FALSE)
}
\arguments{
\item{plan_tbl}{A tibble containing at least \code{original_path}, \code{new_name},
and \code{new_path}.}

\item{overwrite}{Logical. Should existing files be overwritten? Default \code{FALSE}.}
}
\value{
The input tibble with an added logical column \code{copied}.
}
\description{
Copies files according to a renaming plan generated by \code{build_rename_plan()}.
}
\keyword{internal}
