% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_metadata.R
\name{generate_plot}
\alias{generate_plot}
\title{Generate Custom Plots from Metadata}
\usage{
generate_plot(
  metadata,
  x,
  y,
  color = NULL,
  facets = NULL,
  filter_expr = NULL,
  breaks = NULL,
  date_labels = NULL,
  angle = 0
)
}
\arguments{
\item{metadata}{A data frame containing the metadata.}

\item{x}{Character. The column name to use for the x-axis.}

\item{y}{Character. The column name to use for the y-axis.}

\item{color}{Character. (Optional) The column name to use for coloring the points/lines.}

\item{facets}{Character. (Optional) The column name for facet wrapping.}

\item{filter_expr}{Expression. (Optional) An expression to filter the metadata.}

\item{breaks}{Character or NULL. (Optional) Breaks for the x-axis (e.g., "1 day", "1 hour").}

\item{date_labels}{Character. (Optional) Date format for x-axis labels.}

\item{angle}{Numeric. (Optional) Angle for x-axis text. Default is 0.}
}
\value{
A ggplot object.
}
\description{
This function generates ggplot visualizations for the provided metadata,
allowing dynamic specification of axes, color, and additional filtering.
}
\examples{
# Línea de tiempo temperatura por cámara
generate_plot(metadata = metadata, x = Image_dttm, y = temp, color = Camera, facets = Camera, filter_expr = File_date <= "2024-08-27")
# Línea de tiempo de luminancia con color
generate_plot(metadata = metadata, x = Image_dttm, y = bLuma, color = Camera, facets = Camera, filter_expr = File_date <= "2024-08-27")
# Puntos de luminancia filtrados por valor mínimo
generate_plot(metadata = metadata, x = Image_dttm, y = bLuma, color = Camera, facets = Camera, filter_expr = File_date <= "2024-08-27" & bLuma >= 250)

}
