% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plot.R
\name{generate_plot}
\alias{generate_plot}
\title{Generate Custom Plots from Metadata}
\usage{
generate_plot(
  processed_metadata,
  x,
  y,
  color = NULL,
  facets = NULL,
  filter_expr = NULL,
  angle = 0
)
}
\arguments{
\item{processed_metadata}{A data frame containing the processed_metadata.}

\item{x}{Character. The column name to use for the x-axis.}

\item{y}{Character. The column name to use for the y-axis.}

\item{color}{Character. (Optional) The column name to use for coloring the points/lines.}

\item{facets}{Character. (Optional) The column name for facet wrapping.}

\item{filter_expr}{Expression. (Optional) An expression to filter the metadata.}

\item{angle}{Numeric. (Optional) Angle for x-axis text. Default is 0.}
}
\value{
A ggplot object.
}
\description{
This function generates ggplot visualizations for the provided metadata,
allowing dynamic specification of axes, color, and additional filtering.
}
\examples{
# Timeline of temperature for each camera
generate_plot(processed_metadata = processed_metadata,
              x = Image_dttm, y = temp,
              color = Camera, facets = Camera,
              filter_expr = File_date <= "2024-08-27")
# Timeline of brightness luminance and color
generate_plot(processed_metadata = processed_metadata,
              x = Image_dttm, y = bLuma,
              color = Camera, facets = Camera,
              filter_expr = File_date <= "2024-08-27")
# Brightness luminance points filtered by minimum values
generate_plot(processed_metadata = processed_metadata,
              x = Image_dttm, y = bLuma,
              color = Camera, facets = Camera,
              filter_expr = File_date <= "2024-08-27" & bLuma >= 250)

}
