% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metadata.R
\name{process_metadata}
\alias{process_metadata}
\title{Process Camera Trap Metadata}
\usage{
process_metadata(output_dir, output_file)
}
\arguments{
\item{output_dir}{Character. Path to the main directory containing camera folders with renamed images.}

\item{output_file}{Character. Experimental. When set to TRUE, it will read and try to parse the info
within the UserComment field. If you run it and the function fails, try with FALSE.}
}
\value{
A tibble with the processed metadata.
}
\description{
Extracts EXIF metadata from images in a directory, processes the 'UserComment' column,
and saves the resulting metadata to a CSV file.
}
\examples{
# Define paths
output_dir <- "~/Descargas/test1"
output_file <- "12_ago_2024_metadata.csv"

# Run the function
metadata <- process_metadata(output_dir, output_file, UserComment = FALSE)

print(metadata)

}
