% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_images.R
\name{process_metadata}
\alias{process_metadata}
\title{Process Camera Trap Metadata}
\usage{
process_metadata(main_dir, output_file)
}
\arguments{
\item{main_dir}{Character. Path to the main directory containing camera folders with renamed images.}

\item{output_file}{Character. Path to the output CSV file where metadata will be saved.}
}
\value{
A tibble with the processed metadata.
}
\description{
Extracts EXIF metadata from images in a directory, processes the 'UserComment' column,
and saves the resulting metadata to a CSV file.
}
\examples{
# Define directorios
main_dir <- "C:/rename_IA_fish_12_26_ago_2024"
output_file <- "12_26_ago_2024_images_metadata.csv"

# Ejecuta la función consolidada
metadata <- process_camera_trap_metadata(main_dir, output_file)

# Revisa los resultados
print(metadata)

}
