% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_training_images.R
\name{process_training_images}
\alias{process_training_images}
\title{Process and Select Images for Training}
\usage{
process_training_images(
  metadata_file,
  dest_dir,
  output_sample_file,
  output_annotation_file,
  sample_proportion = 0.5,
  filter_by,
  min_lum = NULL,
  max_lum = NULL,
  min_per_dir = 1,
  max_per_dir = Inf
)
}
\arguments{
\item{metadata_file}{Character. Path to the CSV file containing metadata.}

\item{dest_dir}{Character. Path to the directory where the sampled images will be copied.}

\item{output_sample_file}{Character. Path to save the sampled images' metadata as a CSV file.}

\item{output_annotation_file}{Character. Path to save the annotation file.}

\item{sample_proportion}{Numeric. Proportion of images to sample (default is 0.5).}

\item{filter_by}{Unquoted column name giving the luminosity variable to be used for filtering
(e.g. \code{bLuma})}

\item{min_lum}{Numeric. Lower luminosity threshold. Images with luminosity below this value are
excluded. If \code{NULL}, the minimum of \code{filter_by} in the data is used.}

\item{max_lum}{Numeric. Upper luminosity threshold. Images with luminosity above this value are
excluded. If \code{NULL}, the maximum of \code{filter_by} in the data is used.}

\item{min_per_dir}{Integer. Minimum number of images to sample per directory (default 1).}

\item{max_per_dir}{Integer. Maximum number of images to sample per directory (default Inf).}
}
\value{
A tibble with the sampled images' metadata.
}
\description{
This function filters metadata to select daytime images for training, samples a subset of images,
copies the sampled images to a destination directory, and generates a CSV file for further annotation.
}
\examples{
# Parameters
metadata_file <- "12_26_ago_2024_images_metadata.csv"
dest_dir <- "C:/train.images_ago_2024"
output_sample_file <- "ago_2024_images_sampled.csv"
output_annotation_file <- "images_ago_2024.csv"

# Example call
sampled_images <- process_training_images(
  metadata_file = metadata_file,
  dest_dir = dest_dir,
  output_sample_file = output_sample_file,
  output_annotation_file = output_annotation_file,
  sample_proportion = 0.5,
  filter_by = bLuma
)

print(sampled_images)

}
