% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Images_roboflow.R
\name{process_training_images}
\alias{process_training_images}
\title{Process and Select Images for Training}
\usage{
process_training_images(
  metadata_file,
  dest_dir,
  output_sample_file,
  output_annotation_file,
  sample_proportion = 0.5,
  date_cutoff = Sys.Date(),
  start_hour = 6,
  end_hour = 18
)
}
\arguments{
\item{metadata_file}{Character. Path to the CSV file containing metadata.}

\item{dest_dir}{Character. Path to the directory where the sampled images will be copied.}

\item{output_sample_file}{Character. Path to save the sampled images' metadata as a CSV file.}

\item{output_annotation_file}{Character. Path to save the annotation file.}

\item{sample_proportion}{Numeric. Proportion of images to sample (default is 0.5).}

\item{date_cutoff}{Date. Maximum date to include images (default is \code{Sys.Date()}).}

\item{start_hour}{Numeric. Start of daytime hours (default is 6).}

\item{end_hour}{Numeric. End of daytime hours (default is 18).}
}
\value{
A tibble with the sampled images' metadata.
}
\description{
This function filters metadata to select daytime images for training, samples a subset of images,
copies the sampled images to a destination directory, and generates a CSV file for further annotation.
}
\examples{
# Define parámetros
metadata_file <- "12_26_ago_2024_images_metadata.csv"
dest_dir <- "C:/train.images_ago_2024"
output_sample_file <- "ago_2024_images_sampled.csv"
output_annotation_file <- "images_ago_2024.csv"

# Ejecutar la función
sampled_images <- process_training_images(
  metadata_file = metadata_file,
  dest_dir = dest_dir,
  output_sample_file = output_sample_file,
  output_annotation_file = output_annotation_file,
  sample_proportion = 0.5,
  date_cutoff = as.Date("2024-08-26"),
  start_hour = 6,
  end_hour = 18
)

# Revisar resultados
print(sampled_images)

}
