% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_files.R
\name{rename_images}
\alias{rename_images}
\title{Rename Images Based on EXIF Data}
\usage{
rename_images(main_dir, output_dir)
}
\arguments{
\item{main_dir}{Character. Path to the main directory containing camera folders.}

\item{output_dir}{Character. Path to the directory where renamed images will be stored.}
}
\value{
A tibble summarizing the renamed files and their paths.
}
\description{
This function processes images stored in subdirectories, extracts EXIF metadata,
and renames them based on their file modification date, camera folder name, and subfolder.
}
\examples{
# Define rutas de entrada y salida
main_dir <- "C:/IA_fish_12_26_ago_2024"
output_dir <- "C:/rename_IA_fish_12_26_ago_2024"

# Ejecuta la función
renamed_files <- rename_images_with_date(main_dir, output_dir)

# Revisión de resultados
print(renamed_files)

}
