% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_images.R
\name{rename_images}
\alias{rename_images}
\title{Rename Images Based on EXIF Data}
\usage{
rename_images(main_dir, output_dir)
}
\arguments{
\item{main_dir}{Character. Path to the main directory containing camera folders.}

\item{output_dir}{Character. Path to the directory where renamed images will be stored.}
}
\value{
A tibble summarizing the renamed files and their paths.
}
\description{
This function processes images stored in subdirectories, extracts EXIF metadata,
and renames them based on their file modification date, camera folder name, and subfolder.
}
\examples{
# Define input and output paths
main_dir <- "~/Descargas/DSCF0099"
output_dir <- "~/Descargas/test1"

# Run the function
renamed_files <- rename_images(main_dir, output_dir)

print(renamed_files)

}
